implementing update costumer and confirm delete dialog

+++ front/src/pages/modules/costumer/Costumers.tsx
@@ -2,18 +2,21 @@ import { useState, useEffect } from "react";
 import { api } from "../../../../services/API";
 import { Notification } from "../../../utils/Notification";
 import { useNavigate } from "react-router-dom";
+import { UpdateCostumer } from "./UpdateCostumer";
+import { formatDate } from "../../../utils/formatDate";
+import Swal from "sweetalert2";
 interface Costumer {
   id: number;
   name: string;
   contact: string;
-  created_at: string;
-  updated_at: string;
+  created_at: Date;
+  updated_at: Date;
 }
 
 export default function Costumers() {
   const [costumers, setCostumers] = useState<Costumer[]>([]);
   const [value, setValue] = useState(0);
-  const [termo, setTermo] = useState('');
+  const [termo, setTermo] = useState("");
 
   const navigate = useNavigate();
   const filterCostumer = () => {
@@ -23,17 +26,38 @@ export default function Costumers() {
   };
 
   const deleteCostumer = async (id: number) => {
-    await api.delete(`costumer/${id}`).then(() => {
-      Notification.fire({
-        icon: "success",
-        title: "Cliente excluído com sucesso!",
-      });
-      setValue((c) => c + 1);
+    const styledModal = Swal.mixin({
+      customClass: {
+        confirmButton: "btn btn-success",
+        cancelButton: "btn btn-danger",
+      },
+      buttonsStyling: true,
     });
+
+    styledModal
+      .fire({
+        title: "Deseja realmente excluir?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Sim",
+        cancelButtonText: "Não",
+        reverseButtons: true,
+      })
+      .then(async (result) => {
+        if (result.isConfirmed) {
+          await api.delete(`costumer/${id}`).then(() => {
+            setValue((c) => c + 1);
+            window.location.reload(), 4000;
+          });
+          styledModal.fire("Cliente excluído com sucesso!", "", "success");
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          styledModal.fire("Operação cancelada!", "", "error");
+        }
+      });
   };
 
   useEffect(() => {
-    api.get(`costumer`).then((res: any) => {
+    api.get(`costumer`).then((res) => {
       setCostumers(res.data);
     });
   }, []);
@@ -56,7 +80,7 @@ export default function Costumers() {
             type="text"
             id="termo"
             value={termo}
-            onChange={e => setTermo(e.target.value)}
+            onChange={(e) => setTermo(e.target.value)}
             className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
             placeholder="digite"
           />
@@ -103,15 +127,14 @@ export default function Costumers() {
                   <td className="py-4 px-6">{costumer.id}</td>
                   <td className="py-4 px-6">{costumer.name}</td>
                   <td className="py-4 px-6">{costumer.contact}</td>
-                  <td className="py-4 px-6">{costumer.created_at}</td>
-                  <td className="py-4 px-6">{costumer.updated_at}</td>
+                  <td className="py-4 px-6">
+                    {formatDate(costumer.created_at)}
+                  </td>
+                  <td className="py-4 px-6">
+                    {formatDate(costumer.updated_at)}
+                  </td>
                   <td className="py-4 px-1">
-                    <button
-                      type="button"
-                      className="text-white bg-blue-700 hover:bg-blue-800 focus:outline-none focus:ring-4 focus:ring-blue-300 font-medium rounded-full text-sm px-5 py-2.5 text-center mr-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
-                    >
-                      Editar
-                    </button>
+                    <UpdateCostumer id={costumer.id} />
                   </td>
                   <td className="py-4 px-1">
                     <button
